% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BibWrangleR.R
\docType{package}
\name{BibWrangleR}
\alias{BibWrangleR}
\alias{BibWrangleR-package}
\title{BibWrangleR: a tool for extracting bibliographic data from online database aggregators}
\description{
BibWrangleR: a tool for extracting bibliographic data from online database aggregators
}

