% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ebscoBWR.R
\name{ebscoBWR.f}
\alias{ebscoBWR.f}
\title{Converts bibliographic text files extracted from EbscoHost into analyzable data frame}
\usage{
ebscoBWR.f(path, rmDuplicates = TRUE, firstAbstractOnly = TRUE,
  csv = FALSE)
}
\arguments{
\item{path}{path directory for the folder containing text files to be wrangled}

\item{rmDuplicates}{logical vector indicating whether the function should identify and remove duplicate article records}

\item{firstAbstractOnly}{locgical vector indicating whether to retain all available abstract fields for a given article record.  See front end notation.}

\item{csv}{logical vector indicating whether the final data frame generated by the function should be saved to the working directory as a .csv file}
}
\description{
Converts bibliographic text files extracted from EbscoHost into analyzable data frame
}
\note{
ebsoBWR.f will wrangle all text files present in the path folder.
}
\examples{
ebscoBWR.f(csv=TRUE, path="C:/Users/JaneDoe/Desktop/EbscoHostTextFiles")
}

