% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proquestBWR.R
\name{proQuestBWR.f}
\alias{proQuestBWR.f}
\title{Converts bibliographic text files extracted from ProQuest into analyzable data frame}
\usage{
proQuestBWR.f(csv = FALSE, path)
}
\arguments{
\item{csv}{- logical vector indicating whether the final data frame generated by the function should be saved to the working directory as a .csv file}

\item{path}{- path for the folder containing text files to be wrangled}
}
\description{
Converts bibliographic text files extracted from ProQuest into analyzable data frame
}
\note{
proQuestBWR.f will wrangle all text files present in the path folder.
}
\examples{
proQuestBWR.f(csv=TRUE, path="C:/Users/JohnDoe/Desktop/ProQuestTextFiles")
}

